+++ R/bfs_get_catalog.R
@@ -9,19 +9,20 @@
 #'
 #' @importFrom tidyRSS tidyfeed
 #' @importFrom janitor clean_names
+#' @importFrom lifecycle deprecate_warn
 #'
 #' @seealso \code{\link{bfs_get_data}}
 #'
 #' @export
 bfs_get_catalog <- function(language = "de", type = "data") {
-  if (missing(language)) stop("must choose a language, either 'de', 'fr', 'it' or 'en'", call. = FALSE)
+  lifecycle::deprecate_warn("0.5.6", "bfs_get_catalog()", details = "Please use `bfs_get_catalog_data()` or 'bfs_get_catalog_tables()' instead")
   language <- match.arg(arg = language, choices = c("de", "fr", "it", "en"))
-  if (missing(type)) stop("must choose a type, either 'data' or 'tables'", call. = FALSE)
   type <- match.arg(arg = type, choices = c("data", "tables"))
   if (type == "data") {
-    BFS::bfs_get_catalog_data(language = language)
+    catalog <- BFS::bfs_get_catalog_data(language = language)
   }
   if (type == "tables") {
-    BFS::bfs_get_catalog_tables(language = language)
+    catalog <- BFS::bfs_get_catalog_tables(language = language)
   }
+  return(catalog)
 }