+++ R/bfs_get_data_comments.R
@@ -4,7 +4,6 @@
 #'
 #' @param number_bfs The BFS number of a dataset.
 #' @param language Language of the dataset to be translated if exists, i.e. "de", "fr", "it" or "en".
-#' @param url_bfs The URL page of a dataset.
 #' @param query a list with named values, a json query file or json query string using \code{pxweb::pxweb_query()}.
 #' @param clean_names Clean column names using \code{janitor::clean_names()}
 #' @param delay Integer Number of seconds to wait before query using \code{Sys.sleep()}.
@@ -17,38 +16,18 @@
 #' @importFrom magrittr %>%
 #'
 #' @export
-bfs_get_data_comments <- function(number_bfs = NULL, language = "de", url_bfs = NULL, query = NULL, clean_names = FALSE, delay = NULL) {
-  if (is.null(number_bfs) && is.null(url_bfs)) {
-    stop("Please fill bfs_number or url_bfs", call. = FALSE)
-  }
-  if (!is.null(number_bfs) && !is.null(url_bfs)) {
-    stop("Please fill only bfs_number or url_bfs", call. = FALSE)
-  }
-
-  if (!is.null(url_bfs) && is.null(number_bfs)) {
-    html_raw <- xml2::read_html(url_bfs)
-    html_table <- rvest::html_node(html_raw, ".table")
-    df_table <- rvest::html_table(html_table)
-    number_bfs <- df_table$X2[grepl("px", df_table$X2)]
-    if (!startsWith(number_bfs, "px")) {
-      stop("Failed to get the bfs number from: ", url_bfs, "\nPlease add manually the bfs number withs bfs_number.", call. = FALSE)
-    }
-    number_bfs
-  }
-
+bfs_get_data_comments <- function(number_bfs, language = "de", query = NULL, clean_names = FALSE, delay = NULL) {
+  language <- match.arg(arg = language, choices = c("de", "fr", "it", "en"))
   pxweb_api_url <- paste0("https://www.pxweb.bfs.admin.ch/api/v1/", language, "/", number_bfs, "/", number_bfs, ".px")
-
   # check if too many requests HTTP 429
   df_json <- httr2::request("https://www.pxweb.bfs.admin.ch/api/v1") %>%
     httr2::req_url_path_append(paste0(language, "/", number_bfs, "/", number_bfs, ".px")) %>%
     httr2::req_retry(max_tries = 2, max_seconds = 10) %>%
     httr2::req_perform() %>%
     httr2::resp_body_json(simplifyVector = TRUE)
-
   if(!is.null(delay)) {
     Sys.sleep(delay) # waiting time in seconds before query
   }
-  
   if (is.null(query)) {
     variables <- df_json$variables$code
     values <- df_json$variables$values
@@ -64,15 +43,12 @@ bfs_get_data_comments <- function(number_bfs = NULL, language = "de", url_bfs = 
     dims <- query
     pxq <- pxweb::pxweb_query(dims)
   }
-
   df_pxweb <- pxweb::pxweb_get(url = pxweb_api_url, query = pxq)
   comments <- pxweb::pxweb_data_comments(df_pxweb)
   df_comments <- as.data.frame(comments)
   tbl <- tibble::as_tibble(df_comments)
-
   if (clean_names) {
     tbl <- janitor::clean_names(tbl)
   }
-
   return(tbl)
 }