+++ tests/testthat/test-bfs_get_catalog.R
@@ -1,0 +1,5 @@
+test_that("bfs_get_catalog() returns a none-empty data.frame", {
+  df_catalog_data <- BFS::bfs_get_catalog()
+  expect_s3_class(df_catalog_data, "data.frame")
+  expect_true(nrow(df_catalog_data) > 1)
+})