+++ tests/testthat/test-bfs_get_metadata.R
@@ -1,0 +1,5 @@
+test_that("bfs_get_metadata() works", {
+  df <- BFS::bfs_get_metadata(number_bfs = "px-x-1502040100_131")
+  expect_s3_class(df, "data.frame")
+  expect_equal(ncol(df), 7)
+})